Add an endpoint to accept schedule lookup requests
Accepts a PUT request to store a schedule, mapping a route (from station
and to station) to a simple range for times during which we wish to search
for that route.

Just stores that schedule in memory for now, and adds a hook to do a
route search for the route, if the schedule is 'active' (i.e. its times
cover 'now').